# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Producto",
        "description": "CRUD de productos",
        "contact": {
            "email": "soporte@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://api-producto.test",
            "description": "Servidor local Laragon"
        }
    ],
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listar productos",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "responses": {
                    "200": {
                        "description": "Listado paginado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductCollection"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear producto",
                "operationId": "785644b7d7dd38af4b51dafd0fbc1b6d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductMultipartInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Mostrar un producto",
                "operationId": "c87fb2e4d617e7b36c712cad358acdb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar un producto",
                "operationId": "416e4959a3919547eb609d87e3db9135",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductMultipartInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar un producto",
                "operationId": "b49ce9f3cb168b233245548fb4ef36d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminado"
                    }
                }
            }
        },
        "/api/ping": {
            "get": {
                "summary": "Ping de prueba",
                "operationId": "fb0041b0ac351c26a48f4e3a154c4675",
                "responses": {
                    "200": {
                        "description": "pong"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "properties": {
                    "id": {
                        "description": "Todas las definiciones de esquema van en **un único** DocBlock\npegado a la clase, así Swagger-PHP las registra sin problemas.",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Camisa"
                    },
                    "description": {
                        "type": "string",
                        "example": "Camisa de algodón"
                    },
                    "image_path": {
                        "type": "string",
                        "example": "products/xyz.jpg",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ProductInput": {
                "required": [
                    "name"
                ],
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Product"
                    },
                    {
                        "properties": {
                            "id": {
                                "readOnly": true
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "ProductCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Product"
                        }
                    },
                    "links": {
                        "type": "object"
                    },
                    "meta": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ProductMultipartInput": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Camisa"
                    },
                    "description": {
                        "type": "string",
                        "example": "Camisa de algodón"
                    },
                    "image": {
                        "description": "Imagen JPG o PNG (máx. 2 MB)",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Productos",
            "description": "CRUD de productos"
        }
    ]
}